fix(contracts): E2 — quick wins de contrato com o backend + CI
Correções pareadas com o backend (auditoria Fase 0; deploy junto com o backend E2):
- Datas: novo helper core/utils/api_date_time.dart (parse ISO-8601 + toLocal) aplicado
  em 7 models — o detalhe do pedido exibia horário UTC (3h adiantado vs lista); reviews,
  perfil público e admin tinham o mesmo padrão sem toLocal
- Quantidade fracionária: ProducerOrderItem.quantity int→double (0.5 kg aparecia como 0
  para o produtor); formatação espelha o lado do cliente (vírgula, 2 casas)
- Telefone: datasource de perfil normaliza com digitsOnly antes do PUT /customers/me
  (backend agora exige ^\d{11}$; a UI continua exibindo com máscara)
- Complemento: opcional nos forms de cadastro e edição (backend nunca exigiu; o form
  forçava o usuário a inventar um)
- Cancelamento: maxLength 1000 no campo detalhes (alinha com @Size do backend)
- Admin: pré-validação de availability no createProducer (mesmo padrão de address/
  paymentMethods) — antes o campo era omitido e o backend devolvia 400 em inglês
- ApiEndpoints: removidos getters mortos (customers, products, order); product(id)
  mantido — a rota GET /products/{id} agora existe no backend (fallback do detalhe
  de produto deixa de dar 404); customerOrderCancel anotado como rota preferencial
  pendente de migração

CI novo (.github/workflows/ci.yml): flutter analyze + flutter test em PR/push para
main/develop — 29 arquivos de teste existiam sem nenhuma execução automática.

Verificação: flutter analyze (0 erros; warnings pré-existentes) + flutter test (151 ok)

+++ lib/features/producer_orders/data/models/producer_order_item_model.dart
@@ -33,7 +33,9 @@ class ProducerOrderItemModel extends ProducerOrderItem {
                   json['total'] as num? ??
                   0)
               .toDouble(),
-      quantity: quantity.toInt(),
+      // O backend modela quantity como BigDecimal e aceita fração (ex.: 0.5 kg);
+      // toInt() truncava 0.5 para 0 na visão do produtor.
+      quantity: quantity.toDouble(),
       unityType:
           json['unityType'] as String? ??
           json['unit'] as String? ??
